Split out ESPColorView into a separate file
